moving db fn to db utils
